Add typing hints & ndarray conversions, fix tests
Add/strengthen type annotations (NDArray, Sequence, ArrayLike) across fronttracking and advection helpers and return signatures; replace many .values uses in tests with .to_numpy() to ensure explicit numpy arrays; add defensive casts (float/np.array) and dtype enforcement (sparse.eye dtype=float) to avoid type issues; guard ShockWave outlet crossing when velocity is None; add existence check for external 'uv' tool in test script; adjust one test tolerance and fix minor test/formatting indentation; add type: ignore comments where necessary to satisfy static typing. These changes aim to improve type-safety and runtime robustness between pandas/NumPy boundaries and to make tests more deterministic.

+++ src/gwtransport/diffusion.py
@@ -406,7 +406,7 @@ def _infiltration_to_extraction_coeff_matrix(
     cumulative_volume_at_cin_tedges = np.concatenate(([0], np.cumsum(infiltration_volume)))
 
     # Compute the cumulative flow at cout_tedges
-    cumulative_volume_at_cout_tedges = np.interp(cout_tedges, tedges, cumulative_volume_at_cin_tedges)
+    cumulative_volume_at_cout_tedges = np.interp(cout_tedges, tedges, cumulative_volume_at_cin_tedges).astype(float)
 
     rt_edges_2d = residence_time(
         flow=flow,
